Add builder that follows the rules (but doesn't work)
I created it including 32 bit libc.so.6, as mentioned... and it doesn't see it... then I added all the necessary dependencies and it doesn't work... so I changed AppRun, the workflow, I tried with `libunionpreload`... and nothing... so what the fuck do you want from me??? Fuck you! The simple version is better than you! Asshole!

PS: if anyone thinks we can do better, you are welcome, otherwise install the 32-bit libraries on the system and use the binary already available in the releases section. Or use Flatpak, at this point, you gain mental health. Dammit!