debloat again
Removes some duplicated and unneeded packages, for example we were pulling several gstreamer packages when steam only has a dependency to one and it already gets pulled when the steam package is installed. 

I also pull the patched bwrap directly from Ajams repo instead of pre-extracting the utils to copy and paste it.